test_project/backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
psycopg[binary]==3.1.19
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
importlib-metadata==8.5.0
Markdown==3.7
peewee==3.17.6
python-status==1.0.1
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
zipp==3.20.2
gunicorn==22.0.0
uvicorn==0.30.1